requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cu111
torchvision==0.9.1+cu111
torchaudio>=0.8.1
opencv-python>=4.5.3.56
pandas>=0.25.3
Pillow>=8.3.1
protobuf>=3.17.3
pycm>=3.2
pydicom>=2.3.0
scikit-image>=0.17.2
scikit-learn>=0.24.2
scipy>=1.5.4
tensorboard>=2.6.0
tensorboard-data-server>=0.6.1
tensorboard-plugin-wit>=1.8.0
timm==0.3.2
tqdm>=4.62.1

einops >= 0.3.0
h5py >= 2.8.0
imageio >= 2.9.0
matplotlib >= 3.3.2

tqdm >= 4.51.0
transformers >= 3.5.1
utils == 1.0.1
Pygments>=2.7.4

opencv-python>=4.5.3.56
pytorch-msssim>=1.0.0
pytorch-ssim>=0.1
toml>=0.10.2

transformers>=4.30.2
huggingface-hub>=0.16.4
PySide6>=6.5.3





